fix: move static context creation to JobHandlerMetrics
